# [Backend] TRT backend & PP-Infer backend support pinned memory