--- a/fastdeploy/runtime.h
+++ b/fastdeploy/runtime.h
@@ -205,14 +205,14 @@ struct FASTDEPLOY_DECL RuntimeOption {
   void SetTrtCacheFile(const std::string& cache_file_path);
 
   /**
-   * @brief Enable pinned memory in trt backend
+   * @brief Enable pinned memory. Pinned memory can be utilized to speedup the data transfer between CPU and GPU. Currently it's only suppurted in TRT backend and Paddle Inference backend.
    */
-  void EnableTrtPinnedMemory();
+  void EnablePinnedMemory();
 
   /**
-   * @brief Disable pinned memory in trt backend
+   * @brief Disable pinned memory
    */
-  void DisableTrtPinnedMemory();
+  void DisablePinnedMemory();
 
   /**
    * @brief Enable to collect shape in paddle trt backend
@@ -231,6 +231,8 @@ struct FASTDEPLOY_DECL RuntimeOption {
   int device_id = 0;
 
   Device device = Device::CPU;
+
+  bool enable_pinned_memory = false;
 
   // ======Only for ORT Backend========
   // -1 means use default value by ort
@@ -268,7 +270,6 @@ struct FASTDEPLOY_DECL RuntimeOption {
   bool trt_enable_int8 = false;
   size_t trt_max_batch_size = 32;
   size_t trt_max_workspace_size = 1 << 30;
-  bool trt_enable_pinned_memory = false;
 
   // ======Only for Poros Backend=======
   bool is_dynamic = false;