@wang-xinyu 2022-10-20
After integrating pinned memory, the inference latency is comparable with trtexec. It mainly optimized the D2H latency.

The latency includes H2D, GPU inference and D2H.

Tested on P40.

The reason why yolov6s has only a little improvement is the output tensor shape of yolov6s(8400x85) is much smaller than yolov5s(2x25200x85) and yolov7(2x25200x85).

|model| before | after | trtexec|
|:--|:--|:--|:--|
|yolov5s | 11.0566ms | 7.34001ms | 6.76131ms|
|yolov6s |  7.17801ms | 6.483ms | 5.69616ms |
|yolov7 | 25.7099ms | 19.9956ms | 19.8108ms |